+++ src/app/chatbot-design-studio/cds-dashboard/cds-canvas/actions/list/cds-action-askgpt-v2/cds-action-askgpt-v2.component.ts
@@ -52,7 +52,7 @@ export class CdsActionAskgptV2Component implements OnInit {
   showAiError: boolean = false;
   searching: boolean = false;
   ai_response: string = "";
-  ai_error: string = "Oops! Something went wrong. Please retry in a few moment."
+  ai_error: string = "";
 
   temp_variables = [];
 
@@ -283,9 +283,9 @@ export class CdsActionAskgptV2Component implements OnInit {
     this.logger.log("action updated: ", this.action)
   }
 
-  onChangeCheckbox(){
+  onChangeCheckbox(target){
     try {
-      this.action.history = !this.action.history;
+      this.action[target] = !this.action[target];
       this.updateAndSaveAction.emit({type: TYPE_UPDATE_ACTION.ACTION, element: this.action});
     } catch (error) {
       this.logger.log("Error: ", error);
@@ -332,6 +332,11 @@ export class CdsActionAskgptV2Component implements OnInit {
           let name = m.slice(2, m.length - 2);
           let attr = this.action.preview.find(v => v.name === name);
 
+          const index = this.temp_variables.findIndex((e) => e.name === name);
+          if(index> -1 ){ //key already exist: do not add it again
+            return;
+          }
+
           if (attr && attr.value) {
             this.temp_variables.push({ name: name, value: attr.value });
 
@@ -370,14 +375,19 @@ export class CdsActionAskgptV2Component implements OnInit {
       element.classList.remove('preview-container-extended')
     }, 200)
 
-    this.openaiService.previewAskPrompt(data).subscribe((ai_response: any) => {
+    this.openaiService.previewAskPrompt(data).subscribe({ next: (ai_response: any)=> {
       this.searching = false;
       setTimeout(() => {
         let element = document.getElementById("preview-container");
         element.classList.add('preview-container-extended')
       }, 200)
+      if(!ai_response.answer){
+        this.showAiError = true;
+        this.ai_error = this.translate.instant('CDSCanvas.AiNoAnswer')
+        return;
+      }
       this.ai_response = ai_response.answer;
-    }, (err) => {
+    }, error: (err)=> {
       this.searching = false;
       this.logger.error("[ACTION GPT-TASK] previewPrompt error: ", err);
       setTimeout(() => {
@@ -389,13 +399,10 @@ export class CdsActionAskgptV2Component implements OnInit {
         this.ai_error = this.translate.instant('CDSCanvas.AiQuotaExceeded')
         return;
       }
-      
-      this.ai_error = this.translate.instant('CDSCanvas.AiError')
-    }, () => {
+    }, complete: ()=> {
       this.logger.debug("[ACTION GPT-TASK] preview prompt *COMPLETE*: ");
       this.searching = false;
-    })
-
+    }});
   }
 
   openAttributesDialog() {