+++ src/app/chatbot-design-studio/cds-dashboard/cds-canvas/actions/list/cds-action-online-agents-v2/cds-action-online-agents.component.ts
@@ -100,6 +100,14 @@ export class CdsActionOnlineAgentsV2Component implements OnInit {
       this.setFormValue();
     }
     this.departments = this.dashboardService.departments
+    
+    //FIX: if chatbot is imported from other env/project --> reset selectedDepartmentId 
+    if(this.action.selectedDepartmentId){
+      let actionDepIndex = this.departments.findIndex(dep => dep._id === this.action.selectedDepartmentId)
+      if(actionDepIndex === -1){
+        this.action.selectedDepartmentId = null;
+      }
+    }
 
   }
   
@@ -179,6 +187,15 @@ export class CdsActionOnlineAgentsV2Component implements OnInit {
       }
     } catch (error) {
       this.logger.log('error: ', error);
+    }
+  }
+  
+  onChangeCheckbox(target){
+    try {
+      this.action[target] = !this.action[target];
+      this.updateAndSaveAction.emit({type: TYPE_UPDATE_ACTION.ACTION, element: this.action});
+    } catch (error) {
+      this.logger.log("Error: ", error);
     }
   }