+++ src/app/modals/network-offline/network-offline.component.ts
@@ -15,7 +15,6 @@ export class NetworkOfflineComponent implements OnInit {
   ) { }
 
   ngOnInit(): void {
-
     this.networkService.networkStatus$.subscribe((isOnline)=> {
       if(isOnline){
         this.dialogRef.close()